[move] TripRecordViewHistory 관련 Controller, Repository, Service triprecord 도메인 패키지 하위로 이동
related to: #140
