+++ src/main/java/com/haejwo/tripcometrue/domain/triprecord/repository/triprecord_viewhistory/TripRecordViewHistoryRepository.java
@@ -1,4 +1,4 @@
-package com.haejwo.tripcometrue.domain.triprecordViewHistory.repository;
+package com.haejwo.tripcometrue.domain.triprecord.repository.triprecord_viewhistory;
 import com.haejwo.tripcometrue.domain.member.entity.Member;
 import com.haejwo.tripcometrue.domain.triprecord.entity.TripRecordViewHistory;
 import java.util.Optional;
