+++ src/main/java/com/haejwo/tripcometrue/domain/triprecord/service/TripRecordViewHistoryService.java
@@ -1,12 +1,12 @@
-package com.haejwo.tripcometrue.domain.triprecordViewHistory.service;
+package com.haejwo.tripcometrue.domain.triprecord.service;
 
 import com.haejwo.tripcometrue.domain.member.entity.Member;
 import com.haejwo.tripcometrue.domain.triprecord.entity.TripRecord;
 import com.haejwo.tripcometrue.domain.triprecord.entity.TripRecordViewHistory;
 import com.haejwo.tripcometrue.domain.triprecord.exception.TripRecordNotFoundException;
 import com.haejwo.tripcometrue.domain.triprecord.repository.triprecord.TripRecordRepository;
-import com.haejwo.tripcometrue.domain.triprecordViewHistory.dto.response.TripRecordViewHistoryResponseDto;
-import com.haejwo.tripcometrue.domain.triprecordViewHistory.repository.TripRecordViewHistoryRepository;
+import com.haejwo.tripcometrue.domain.triprecord.dto.response.triprecord_viewhistory.TripRecordViewHistoryResponseDto;
+import com.haejwo.tripcometrue.domain.triprecord.repository.triprecord_viewhistory.TripRecordViewHistoryRepository;
 import com.haejwo.tripcometrue.global.springsecurity.PrincipalDetails;
 import jakarta.transaction.Transactional;
 import java.util.Optional;
@@ -53,6 +53,5 @@ public class TripRecordViewHistoryService {
     return tripRecordViewHistoryRepository.findByMember(principalDetails.getMember(), pageable)
         .map(TripRecordViewHistoryResponseDto::fromEntity);
   }
-
 
 }
